Module ra_hir_expand::db

Defines database & queries for macro expansion.

Structs

AstDatabaseGroupStorage__

AstDatabaseStorage
Representative struct for the query group.

AstIdMapQuery

InternEagerExpansionLookupQuery

InternEagerExpansionQuery

InternMacroLookupQuery

InternMacroQuery

MacroArgQuery

MacroDefQuery

MacroExpandQuery

ParseMacroQuery

Enums

AstDatabaseGroupKey__

TokenExpander

Traits

AstDatabase

Functions

expand_hypothetical
This expands the given macro call, but with different arguments. This is used for completion, where we want to see what 'would happen' if we insert a token. The token_to_map mapped down into the expansion, with the mapped token returned.

parse_macro_with_arg