Module ra_hir_expand::name

FIXME: write short doc here

Structs

Name
Name is a wrapper around string, which is used in hir for both references and declarations. In theory, names should also carry hygiene info, but we are not there yet!

Constants

BOOL

BOXED

BOX_TYPE

CHAR

F32

F64

FUTURE

FUTURE_TYPE

I8

I16

I32

I64

I128

INTO_ITERATOR_TYPE

ISIZE

ITEM_TYPE

ITER

MACRO_RULES

OK_TYPE

OPS

OUTPUT_TYPE

RESULT

RESULT_TYPE

SELF_PARAM

SELF_TYPE

STD

STR

TARGET_TYPE

TRY_TYPE

U8

U16

U32

U64

U128

USIZE

Traits

AsName